move propagation to its own package
